refactor(ImagePreview): add ImagePreviewItem
