+++ src/image-preview/shared.js
@@ -1,0 +1,5 @@
+import { createNamespace } from '../utils';
+
+const [createComponent, bem] = createNamespace('image-preview');
+
+export { createComponent, bem };
